requirements.txt
streamlit
pandas
numpy==1.24.4
scikit-learn==1.4.2
matplotlib
seaborn
plotly
python-dateutil
pycaret
streamlit-option-menu